# Conflicts:
#	WhirlyGlobeSrc/AutoTester/AutoTester.xcodeproj/project.pbxproj
#	WhirlyGlobeSrc/AutoTester/AutoTester.xcworkspace/xcshareddata/AutoTester.xcscmblueprint
#	WhirlyGlobeSrc/AutoTester/AutoTester/testCases/ScreenLabelsTestCase.swift
#	WhirlyGlobeSrc/AutoTester/Podfile
#	WhirlyGlobeSrc/AutoTester/Podfile.lock
#	WhirlyGlobeSrc/AutoTester/Pods/Manifest.lock
#	WhirlyGlobeSrc/AutoTester/Pods/Pods.xcodeproj/project.pbxproj
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Alamofire/Alamofire.xcconfig
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Alamofire/Info.plist
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Pods-AutoTester/Pods-AutoTester-acknowledgements.markdown
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Pods-AutoTester/Pods-AutoTester-acknowledgements.plist
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Pods-AutoTester/Pods-AutoTester-frameworks.sh
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Pods-AutoTester/Pods-AutoTester.debug.xcconfig
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/Pods-AutoTester/Pods-AutoTester.release.xcconfig
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/SwiftyDropbox/Info.plist
#	WhirlyGlobeSrc/AutoTester/Pods/Target Support Files/SwiftyDropbox/SwiftyDropbox.xcconfig


Former-commit-id: f105c39bfe29ea473d87498d986d97990c1c7f41

+++ WhirlyGlobeSrc/WhirlyGlobe-MaplyComponent/src/MaplyGeomModel.mm
@@ -78,7 +78,7 @@ using namespace Eigen;
     {
         if (geom.texId >= 0 && geom.texId < texFileMap.size())
         {
-            geom.texId = texFileMap[geom.texId];
+            geom.texId = (int)texFileMap[geom.texId];
         }
         outRawGeom.push_back(geom);
     }
@@ -155,12 +155,12 @@ using namespace Eigen;
                 {
                     stringGeom[rect.subTex.texId] = GeometryRaw();
                     geom = &stringGeom[rect.subTex.texId];
-                    geom->texId = rect.subTex.texId;
+                    geom->texId = (int)rect.subTex.texId;
                 } else
                     geom = &stringGeom[rect.subTex.texId];
 
                 // Convert and transform the points
-                int basePt = geom->pts.size();
+                int basePt = (int)geom->pts.size();
                 geom->pts.reserve(geom->pts.size()+4);
                 Point2d pts[4];
                 pts[0] = Point2d(rect.pts[0].x(),rect.pts[0].y());
